* fixed errors reported by gcc 4.x / generic error * fixed xml character outfit saving

+++ otserv/branches/revbattlesys/ioplayerxml.cpp
@@ -7,7 +7,7 @@
 // modify it under the terms of the GNU General Public License
 // as published by the Free Software Foundation; either version 2
 // of the License, or (at your option) any later version.
-// 
+//
 // This program is distributed in the hope that it will be useful,
 // but WITHOUT ANY WARRANTY; without even the implied warranty of
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
@@ -79,7 +79,7 @@ bool IOPlayerXML::loadPlayer(Player* player, std::string name)
 		xmlMutexUnlock(xmlmutex);
 		Account a = IOAccount::instance()->loadAccount(account);
 		xmlMutexLock(xmlmutex);
-		
+
 		player->password = a.password;
 		if(a.accnumber == 0 || a.accnumber != (unsigned long)account){
 		  xmlFreeDoc(doc);
@@ -150,22 +150,22 @@ bool IOPlayerXML::loadPlayer(Player* player, std::string name)
 			isLoaded = false;
 
 		player->setNormalSpeed();
-		
+
 		if(readXMLInteger(root, "lastlogin", intValue)){
 			player->lastLoginSaved = intValue;
 		}
 		else
 			player->lastLoginSaved = 0;
-		
+
 		//level percent
-		player->level_percent = (unsigned char)(100*(player->experience - player->getExpForLv(player->level)) / 
+		player->level_percent = (unsigned char)(100*(player->experience - player->getExpForLv(player->level)) /
 			(1.*player->getExpForLv(player->level+1) - player->getExpForLv(player->level)));
-		
+
 		p = root->children;
 
 		while(p){
 			if(xmlStrcmp(p->name, (const xmlChar*)"mana") == 0){
-				
+
 				if(readXMLInteger(p, "now", intValue)){
 					player->mana = intValue;
 				}
@@ -183,10 +183,10 @@ bool IOPlayerXML::loadPlayer(Player* player, std::string name)
 				}
 				else
 					isLoaded = false;
-					
+
 			}
 			else if(xmlStrcmp(p->name, (const xmlChar*)"health") == 0){
-				
+
 				if(readXMLInteger(p, "now", intValue)){
 					player->health = intValue;
 
@@ -195,7 +195,7 @@ bool IOPlayerXML::loadPlayer(Player* player, std::string name)
 				}
 				else
 					isLoaded = false;
-				
+
 				if(readXMLInteger(p, "max", intValue)){
 					player->healthMax = intValue;
 				}
@@ -217,7 +217,7 @@ bool IOPlayerXML::loadPlayer(Player* player, std::string name)
 					isLoaded = false;
 
 				player->lookMaster = player->lookType;
-				
+
 				if(readXMLInteger(p, "head", intValue)){
 					player->lookHead = intValue;
 				}
@@ -263,7 +263,7 @@ bool IOPlayerXML::loadPlayer(Player* player, std::string name)
 					isLoaded = false;
 			}
 			else if(xmlStrcmp(p->name, (const xmlChar*)"temple") == 0){
-				
+
 				if(readXMLInteger(p, "x", intValue)){
 					player->masterPos.x = intValue;
 				}
@@ -283,25 +283,25 @@ bool IOPlayerXML::loadPlayer(Player* player, std::string name)
 					isLoaded = false;
 			}
 			else if(xmlStrcmp(p->name, (const xmlChar*)"guild") == 0){
-				
+
 				if(readXMLString(p, "name", strValue)){
 					player->guildName = strValue;
 				}
 				else
 					isLoaded = false;
-				
+
 				if(readXMLString(p, "rank", strValue)){
 					player->guildRank = strValue;
 				}
 				else
 					isLoaded = false;
-				
+
 				if(readXMLString(p, "nick", strValue)){
 					player->guildNick = strValue;
 				}
 				else
 					isLoaded = false;
-				
+
 				if(readXMLInteger(p, "id", intValue)){
 					player->guildId = intValue;
 				}
@@ -315,7 +315,7 @@ bool IOPlayerXML::loadPlayer(Player* player, std::string name)
 					int s_lvl = 0;
 					int s_tries = 0;
 					if(xmlStrcmp(tmpNode->name, (const xmlChar*)"skill") == 0){
-						
+
 						if(readXMLInteger(tmpNode, "skillid", intValue)){
 							s_id = intValue;
 						}
@@ -347,7 +347,7 @@ bool IOPlayerXML::loadPlayer(Player* player, std::string name)
 					if(xmlStrcmp(slotNode->name, (const xmlChar*)"slot") == 0){
 						int itemId = 0;
 						int slotId = 0;
-						
+
 						if(readXMLInteger(slotNode, "slotid", intValue)){
 							slotId = intValue;
 						}
@@ -376,7 +376,7 @@ bool IOPlayerXML::loadPlayer(Player* player, std::string name)
 					if(xmlStrcmp(tmpNode->name, (const xmlChar*)"depot") == 0){
 						int depotId = 0;
 						int itemId = 0;
-						
+
 						if(readXMLInteger(tmpNode, "depotid", intValue)){
 							depotId = intValue;
 						}
@@ -430,11 +430,11 @@ bool IOPlayerXML::loadPlayer(Player* player, std::string name)
 
 		std::cout << "loaded " << filename << std::endl;
 		xmlFreeDoc(doc);
-		xmlMutexUnlock(xmlmutex);		
+		xmlMutexUnlock(xmlmutex);
 		return isLoaded;
 	}
 
-	xmlMutexUnlock(xmlmutex);	
+	xmlMutexUnlock(xmlmutex);
 	return false;
 }
 
@@ -445,8 +445,8 @@ bool IOPlayerXML::savePlayer(Player* player)
 	toLowerCaseString(filename); //store all player files in lowercase
 
 	std::stringstream sb;
-    
-	xmlMutexLock(xmlmutex);    
+
+	xmlMutexLock(xmlmutex);
 	xmlNodePtr nn, sn, pn, root;
 
 	xmlDocPtr doc = xmlNewDoc((const xmlChar*)"1.0");
@@ -458,7 +458,7 @@ bool IOPlayerXML::savePlayer(Player* player)
 	sb << player->getName();  	       xmlSetProp(root, (const xmlChar*)"name", (const xmlChar*)sb.str().c_str());     sb.str("");
 	sb << player->getGUID();           xmlSetProp(root, (const xmlChar*)"id", (const xmlChar*)sb.str().c_str());	sb.str("");
 	sb << player->accountNumber;       xmlSetProp(root, (const xmlChar*)"account", (const xmlChar*)sb.str().c_str());	sb.str("");
-	sb << player->sex;                 xmlSetProp(root, (const xmlChar*)"sex", (const xmlChar*)sb.str().c_str());     	sb.str("");	
+	sb << player->sex;                 xmlSetProp(root, (const xmlChar*)"sex", (const xmlChar*)sb.str().c_str());     	sb.str("");
 	sb << player->getDirection();
 
 	if (sb.str() == "North"){sb.str(""); sb << "0";}
@@ -467,10 +467,10 @@ bool IOPlayerXML::savePlayer(Player* player)
 	if (sb.str() == "West") {sb.str(""); sb << "3";}
 	xmlSetProp(root, (const xmlChar*)"lookdir", (const xmlChar*)sb.str().c_str());  sb.str("");
 
-	sb << player->experience;         xmlSetProp(root, (const xmlChar*)"exp", (const xmlChar*)sb.str().c_str());       sb.str("");	
+	sb << player->experience;         xmlSetProp(root, (const xmlChar*)"exp", (const xmlChar*)sb.str().c_str());       sb.str("");
 	sb << (int)player->getVocationId();      xmlSetProp(root, (const xmlChar*)"voc", (const xmlChar*)sb.str().c_str());       sb.str("");
-	sb << player->level;              xmlSetProp(root, (const xmlChar*)"level", (const xmlChar*)sb.str().c_str());     sb.str("");	
-	sb << player->accessLevel;             xmlSetProp(root, (const xmlChar*)"access", (const xmlChar*)sb.str().c_str());	sb.str("");	
+	sb << player->level;              xmlSetProp(root, (const xmlChar*)"level", (const xmlChar*)sb.str().c_str());     sb.str("");
+	sb << player->accessLevel;             xmlSetProp(root, (const xmlChar*)"access", (const xmlChar*)sb.str().c_str());	sb.str("");
 	sb << player->getCapacity();      xmlSetProp(root, (const xmlChar*)"cap", (const xmlChar*)sb.str().c_str());       sb.str("");
 	sb << player->magLevel;	          xmlSetProp(root, (const xmlChar*)"magLevel", (const xmlChar*)sb.str().c_str());  sb.str("");
 	sb << player->lastlogin;	        xmlSetProp(root, (const xmlChar*)"lastlogin", (const xmlChar*)sb.str().c_str());  sb.str("");
@@ -480,42 +480,43 @@ bool IOPlayerXML::savePlayer(Player* player)
 	sb << player->getLoginPosition().y;  	xmlSetProp(pn,  (const xmlChar*)"y", (const xmlChar*)sb.str().c_str());        sb.str("");
 	sb << player->getLoginPosition().z; 	xmlSetProp(pn,  (const xmlChar*)"z", (const xmlChar*)sb.str().c_str());	       sb.str("");
 	xmlAddChild(root, pn);
-	
+
 	pn = xmlNewNode(NULL,(const xmlChar*)"temple");
 	sb << player->masterPos.x;  xmlSetProp(pn, (const xmlChar*)"x", (const xmlChar*)sb.str().c_str());        sb.str("");
 	sb << player->masterPos.y;  xmlSetProp(pn, (const xmlChar*)"y", (const xmlChar*)sb.str().c_str());        sb.str("");
 	sb << player->masterPos.z; 	xmlSetProp(pn, (const xmlChar*)"z", (const xmlChar*)sb.str().c_str());	       sb.str("");
 	xmlAddChild(root, pn);
-	
+
 	pn = xmlNewNode(NULL,(const xmlChar*)"health");
 	sb << player->health;     xmlSetProp(pn, (const xmlChar*)"now", (const xmlChar*)sb.str().c_str());        sb.str("");
 	sb << player->healthMax;  xmlSetProp(pn, (const xmlChar*)"max", (const xmlChar*)sb.str().c_str());        sb.str("");
 	sb << player->food;  	    xmlSetProp(pn, (const xmlChar*)"food", (const xmlChar*)sb.str().c_str());       sb.str("");
-	                     
+
 	xmlAddChild(root, pn);
-	
+
 	pn = xmlNewNode(NULL,(const xmlChar*)"mana");
 	sb << player->mana;      xmlSetProp(pn, (const xmlChar*)"now", (const xmlChar*)sb.str().c_str());        sb.str("");
 	sb << player->manaMax;   xmlSetProp(pn, (const xmlChar*)"max", (const xmlChar*)sb.str().c_str());        sb.str("");
 	sb << player->manaSpent; xmlSetProp(pn, (const xmlChar*)"spent", (const xmlChar*)sb.str().c_str());      sb.str("");
 	xmlAddChild(root, pn);
-    	               
+
+    //upconversion of uchar(uint8_t) to get value not character into the stream
 	pn = xmlNewNode(NULL,(const xmlChar*)"look");
-	sb << player->lookMaster;  xmlSetProp(pn, (const xmlChar*)"type", (const xmlChar*)sb.str().c_str());        sb.str("");
-	sb << player->lookHead;    xmlSetProp(pn, (const xmlChar*)"head", (const xmlChar*)sb.str().c_str());        sb.str("");
-	sb << player->lookBody;    xmlSetProp(pn, (const xmlChar*)"body", (const xmlChar*)sb.str().c_str());        sb.str("");
-	sb << player->lookLegs;    xmlSetProp(pn, (const xmlChar*)"legs", (const xmlChar*)sb.str().c_str());        sb.str("");
-	sb << player->lookFeet;    xmlSetProp(pn, (const xmlChar*)"feet", (const xmlChar*)sb.str().c_str());        sb.str("");
+	sb << (int16_t)player->lookMaster;  xmlSetProp(pn, (const xmlChar*)"type", (const xmlChar*)sb.str().c_str());        sb.str("");
+	sb << (int16_t)player->lookHead;    xmlSetProp(pn, (const xmlChar*)"head", (const xmlChar*)sb.str().c_str());        sb.str("");
+	sb << (int16_t)player->lookBody;    xmlSetProp(pn, (const xmlChar*)"body", (const xmlChar*)sb.str().c_str());        sb.str("");
+	sb << (int16_t)player->lookLegs;    xmlSetProp(pn, (const xmlChar*)"legs", (const xmlChar*)sb.str().c_str());        sb.str("");
+	sb << (int16_t)player->lookFeet;    xmlSetProp(pn, (const xmlChar*)"feet", (const xmlChar*)sb.str().c_str());        sb.str("");
 	xmlAddChild(root, pn);
-    
+
 	pn = xmlNewNode(NULL,(const xmlChar*)"guild");
 	sb << player->guildName;  xmlSetProp(pn, (const xmlChar*)"name", (const xmlChar*)sb.str().c_str());        sb.str("");
 	sb << player->guildRank;  xmlSetProp(pn, (const xmlChar*)"rank", (const xmlChar*)sb.str().c_str());        sb.str("");
 	sb << player->guildNick;  xmlSetProp(pn, (const xmlChar*)"nick", (const xmlChar*)sb.str().c_str());       sb.str("");
 	sb << player->guildId;  	xmlSetProp(pn, (const xmlChar*)"id",   (const xmlChar*)sb.str().c_str());       sb.str("");
-	
+
 	xmlAddChild(root, pn);
-    	      
+
 	sn = xmlNewNode(NULL,(const xmlChar*)"skills");
 	for(int i = 0; i <= 6; i++){
 	  pn = xmlNewNode(NULL,(const xmlChar*)"skill");
@@ -524,9 +525,9 @@ bool IOPlayerXML::savePlayer(Player* player)
 	  sb << player->skills[i][SKILL_TRIES];     xmlSetProp(pn, (const xmlChar*)"tries",   (const xmlChar*)sb.str().c_str());        sb.str("");
 	  xmlAddChild(sn, pn);
 	}
-	
+
 	xmlAddChild(root, sn);
-	
+
 	sn = xmlNewNode(NULL,(const xmlChar*)"inventory");
 	for(int i = 1; i <= 10; i++){
 		if(player->items[i]){
@@ -534,47 +535,47 @@ bool IOPlayerXML::savePlayer(Player* player)
 			sb << i;
 			xmlSetProp(pn, (const xmlChar*)"slotid", (const xmlChar*)sb.str().c_str());
 			sb.str("");
-          
+
 			nn = player->items[i]->serialize();
 
 			xmlAddChild(pn, nn);
 			xmlAddChild(sn, pn);
 		}
 	}
-	
+
 	xmlAddChild(root, sn);
-	
+
 	sn = xmlNewNode(NULL,(const xmlChar*)"depots");
-	
+
 	for(DepotMap::reverse_iterator it = player->depots.rbegin(); it !=player->depots.rend()  ;++it){
 		pn = xmlNewNode(NULL,(const xmlChar*)"depot");
 		sb << it->first;
 		xmlSetProp(pn, (const xmlChar*)"depotid", (const xmlChar*)sb.str().c_str());
 		sb.str("");
-          
+
 		nn = (it->second)->serialize();
 		xmlAddChild(pn, nn);
 		xmlAddChild(sn, pn);
 	}
-      
+
 	xmlAddChild(root, sn);
-   
+
 	sn = xmlNewNode(NULL,(const xmlChar*)"storage");
 	for(StorageMap::const_iterator cit = player->getStorageIteratorBegin(); cit != player->getStorageIteratorEnd();cit++){
 		pn = xmlNewNode(NULL,(const xmlChar*)"data");
 		sb << cit->first;
 		xmlSetProp(pn, (const xmlChar*)"key", (const xmlChar*)sb.str().c_str());
 		sb.str("");
-          
+
 		sb << cit->second;
 		xmlSetProp(pn, (const xmlChar*)"value", (const xmlChar*)sb.str().c_str());
 		sb.str("");
-        
+
 		xmlAddChild(sn, pn);
 	}
-	
+
 	xmlAddChild(root, sn);
-	
+
 	//Save the character
 	if(xmlSaveFile(filename.c_str(), doc)){
 		#ifdef __DEBUG__
@@ -587,8 +588,8 @@ bool IOPlayerXML::savePlayer(Player* player)
 	}
 	else{
 		std::cout << "\tCouldn't save character =(\n";
-		xmlFreeDoc(doc);       
-		xmlMutexUnlock(xmlmutex);       
+		xmlFreeDoc(doc);
+		xmlMutexUnlock(xmlmutex);
 		return false;
 	}
 }
